+++ src/components/index.js
@@ -3,3 +3,5 @@ export Button from './Button';
 export ColorSelector from './ColorSelector';
 export Dropdown from './Dropdown';
 export OverflowMenu from './OverflowMenu';
+export Icon from './Icon';
+export Table from './Table';